fix: fix get token flow
